# fix(activity): limit the number of logs displayed