--- a/packages/webapp/src/pages/Activity.tsx
+++ b/packages/webapp/src/pages/Activity.tsx
@@ -816,6 +816,9 @@ export default function Activity() {
                                             {activity.id === expandedRow && activity.messages && activity.messages[0] && (
                                                 <>
                                                 <div className="flex flex-col space-y-4 mt-6 font-mono">
+                                                    {activity.messages.length >= 1 && (
+                                                        <div className='text-center text-gray-500'>[only showing the last 1000 logs]</div>
+                                                    )}
                                                     {activity.messages.map((message, index: number) => (
                                                         <div key={index} className="flex flex-col max-w-7xl">
                                                             <div className="whitespace-normal break-all overflow-wrap">