--- a/packages/webapp/src/types.ts
+++ b/packages/webapp/src/types.ts
@@ -7,9 +7,7 @@ export interface ActivityResponse {
     start: number;
     end: number;
     message: string;
-    messages: {
-        [index: string]: undefined | string | number;
-    }[];
+    messages: Record<string, any>[];
     connection_id: string;
     provider_config_key: string;
     provider: string;
@@ -18,9 +16,9 @@ export interface ActivityResponse {
     operation_name?: string;
 }
 
-export interface SyncResult {
-    [key: string]: Result;
-}
+export type ActivityMessageResponse = Record<number, Record<string, any>[]>;
+
+export type SyncResult = Record<string, Result>;
 
 export interface Result {
     added: number;
@@ -44,9 +42,7 @@ export interface Sync {
     connections:
         | {
               connection_id: string;
-              metadata?: {
-                  [key: string]: string | Record<string, string>;
-              };
+              metadata?: Record<string, string | Record<string, string>>;
           }[]
         | null;
     metadata?: {