## Describe your changes

Some jobs can contains 100K+ messages which is not playing well with the API nor the UI.
This is a very short term solution to stop crashing our servers (and customers UI when the request do come).


Right now the query is getting "N activity with N messages" which is cumbersome to achieve in Postgres.

I feel like a better solution (without changing too much things) would be to have one call per Activity. That would make the query far more straightforward, with the compromise of making more queries in the UI. Since this fix would have changed a lot the UI logic, and there is redesign on its way, I didn't wanted to go this way.

## Issue ticket number and link

Fixes NAN-366


## Explain

**old**
```sql
EXPLAIN ANALYZE SELECT
	activity_log_id,
	array_agg(row_to_json(_nango_activity_log_messages.*)) AS messages
FROM
	nango._nango_activity_log_messages
WHERE
	activity_log_id IN(41486578)
	AND environment_id = 1770
GROUP BY
	activity_log_id;

QUERY PLAN
GroupAggregate  (cost=88897.96..151974.19 rows=31997 width=36) (actual time=4797.194..4797.198 rows=1 loops=1)
  Group Key: activity_log_id
  ->  Bitmap Heap Scan on _nango_activity_log_messages  (cost=88897.96..151146.32 rows=57054 width=612) (actual time=682.302..1772.079 rows=501752 loops=1)
        Recheck Cond: ((activity_log_id = 41486578) AND (environment_id = 1770))
        Heap Blocks: exact=120294
        ->  BitmapAnd  (cost=88897.96..88897.96 rows=57054 width=0) (actual time=642.040..642.042 rows=0 loops=1)
              ->  Bitmap Index Scan on activity_log_id_index  (cost=0.00..4837.19 rows=519777 width=0) (actual time=32.992..32.992 rows=501752 loops=1)
                    Index Cond: (activity_log_id = 41486578)
              ->  Bitmap Index Scan on _nango_activity_log_messages_environment_id_index  (cost=0.00..84031.99 rows=8914723 width=0) (actual time=605.274..605.274 rows=10157984 loops=1)
                    Index Cond: (environment_id = 1770)
Planning Time: 0.114 ms
Execution Time: 4836.086 ms
```

**new**
```sql

QUERY PLAN
--
EXPLAIN ANALYZE SELECT
	*
FROM (
	SELECT
		*,
		RANK() OVER (PARTITION BY activity_log_id ORDER BY created_at DESC) AS rank
	FROM
		_nango_activity_log_messages) AS partition
WHERE
	activity_log_id IN(41486578)
	AND environment_id = 1770
	AND partition.rank <= 1000;

QUERY PLAN
Subquery Scan on partition  (cost=535615.77..611745.68 rows=57054 width=592) (actual time=302.647..581.285 rows=1000 loops=1)
  Filter: (partition.environment_id = 1770)
  ->  WindowAgg  (cost=535615.77..605248.46 rows=519777 width=592) (actual time=302.646..581.148 rows=1000 loops=1)
        Run Condition: (rank() OVER (?) <= 1000)
        ->  Gather Merge  (cost=535615.77..596152.37 rows=519777 width=584) (actual time=302.632..540.779 rows=501752 loops=1)
              Workers Planned: 2
              Workers Launched: 2
              ->  Sort  (cost=534615.74..535157.18 rows=216574 width=584) (actual time=294.606..331.427 rows=167251 loops=3)
                    Sort Key: _nango_activity_log_messages.created_at DESC
                    Sort Method: external merge  Disk: 28168kB
                    Worker 0:  Sort Method: external merge  Disk: 19480kB
                    Worker 1:  Sort Method: external merge  Disk: 19528kB
                    ->  Parallel Index Scan using activity_log_id_index on _nango_activity_log_messages  (cost=0.57..482470.71 rows=216574 width=584) (actual time=0.056..164.327 rows=167251 loops=3)
                          Index Cond: (activity_log_id = 41486578)
Planning Time: 0.192 ms
Execution Time: 588.752 ms
```

The new query plan is slightly more complicated but results in less row and less computation, so faster (for this big activity with 500K messages).