@bodinsamuel 2024-02-14
> Just confirming that this doesn't change the ordering of logs at all but just for high volume logs it truncates the beginning set, correct?

Yes, I tested on the UI on my side but I would love double confirmation with what you have in your DB. Might help 👍🏻 

> We should let the user know in some way that the logs were truncated.

I didn't wanted to change the UI too much before the redesign but fixed 👌🏻