@bodinsamuel 2024-02-15
it's weird I never have those errors on my end, good catch 🙏🏻