@khaliqgant 2024-02-15
> > @khaliqgant can you actually double check on your side the order of the logs in the UI? I reverted my fix because it was correct on my side, maybe you were not up to date on the compilation
> 
> Will check in the next couple of hours.

Ordering looks good!